Nice design on the landing page
Glad to see so many improvements for this project. I really like the dynamic background,  together with the color palette that matches the theme perfectly. I would suggest adding more details on the top of the landing page since there are only the "sign-in/up" on the top right, and the left side looks kind of empty.